Working idea of depending on the library by the services
