+++ discoverable-client/build.gradle
@@ -48,6 +48,9 @@ gitProperties {
 }
 
 dependencies {
+    implementation(project(':mock-zos'))
+    testImplementation(project(':mock-zos'))
+    
     implementation project(':onboarding-enabler-spring')
     implementation project(':zaas-client')
     implementation libraries.springFox
