+++ gateway-service/build.gradle
@@ -57,6 +57,9 @@ configurations.all {
 }
 
 dependencies {
+    compileOnly(project(':mock-zos'))
+    testImplementation(project(':mock-zos'))
+    
     implementation(project(':apiml-security-common'))
     implementation libraries.spring_boot_starter_webflux
     implementation libraries.spring_cloud_starter_zuul
