+++ mock-zos/build.gradle
@@ -1,4 +1,13 @@
 dependencies {
+    implementation libraries.lombok
+    implementation libraries.apache_commons_text
+    
+    annotationProcessor libraries.lombok
+
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+}
+
+jar {
+    baseName = "mock-zos"
 }
