--- mock-zos/src/main/java/com/ibm/jzos/ZUtil/TestClass.java
@@ -1,17 +1,0 @@
-/*
- * This program and the accompanying materials are made available under the terms of the
- * Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Copyright Contributors to the Zowe Project.
- */
-package com.ibm.jzos.ZUtil;
-
-public abstract class TestClass implements ZUtil {
-    @Override
-    public String getCurrentJobId() {
-        return null;
-    }
-}
